requirements.txt
deepface==0.0.98
fastapi==0.128.2
fastmcp==2.14.5
httpx==0.28.1
numpy==2.4.2
openai_whisper==20250625
opencv_contrib_python==4.13.0.90
opencv_python==4.12.0.88
pydantic_ai==1.56.0
python-dotenv==1.2.1
pyttsx3==2.99
reachy_mini==1.3.1
scipy==1.17.0
silero_vad==6.2.0
soundfile==0.13.1
torch==2.10.0
torchaudio==2.10.0
transformers==5.0.0
uvicorn==0.40.0
